fix(trails): unbox Double before narrowing cast to int
NumberSetting.get() returns Double. Java casting rules only allow
unboxing followed by *widening* primitive conversion in a single cast,
so '(int) length.get()' fails to compile with:

    error: incompatible types: Double cannot be converted to int

Use the same pattern the rest of the codebase uses (see TargetESP):
explicit .doubleValue() unbox first, then primitive narrowing cast.

+++ src/main/java/dev/blumdlc/client/modules/impl/Trails.java
@@ -130,7 +130,7 @@ public final class Trails extends Module {
 		samples.addLast(new Sample(pos.x, pos.y, pos.z, player.getHeight(), now));
 		lastSampleMs = now;
 
-		int maxSamples = Math.max(2, (int) length.get());
+		int maxSamples = Math.max(2, (int) length.get().doubleValue());
 		while (samples.size() > maxSamples) {
 			samples.pollFirst();
 		}
